pyproject.toml
[project]
name = "nanodjango-fileapi"
version = "0.1.0"
description = "Simple token authenticated API for uploading and downloading files."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nanodjango>=0.9.1",
    "django-ninja>=1.3.0",
    "environs>=11.0.0",
]
